Fixes broken use of url exceptions when using the decorator
Up to now using url exceptions when using the decorator and not the middleware
has been broken. This commit fixes that behavior and adds proper tests to check
that.

Using url exceptions in the decorator should be a very, very rare case anyway,
because all it can be used for is locking access to a view via a set of urls,
but allowing it via another set.